Updated test and refactored handling of screenshots
